requirements.txt
ultralytics>=8.0.0
torch>=1.13.0
opencv-python
numpy
matplotlib
pyyaml
streamlit
Pillow
